+++ CHANGELOG.md
@@ -4,9 +4,14 @@
 
 ## Master
 
+### New Features
+
+- Added support in `AutoHashable` for static variables, `[Hashable]` array and `[Hashable: Hashable]` dictionary
+
 ### Bug fixes
 
 - Fixed linker issue when using Swift templates
+
 
 ## 0.7.2